Refactor : rename MainModel.js to CalculatorModel.js
